feat/將物件進行簡單拆分，新增configs資料夾
