--- .env
@@ -1,7 +1,0 @@
-# 您的 Figma Personal Access Token
-# 您可以從 Figma 的 "Help and account" > "Account settings" > "Personal access tokens" 建立一個。
-FIGMA_API_KEY="YOUR_FIGMA_API_KEY"
-
-# 您的 Google AI API Key
-# 您可以從 Google AI Studio (https://aistudio.google.com/) 取得。
-GOOGLE_API_KEY="YOUR_GOOGLE_API_KEY"
