--- GEMINI.MD
@@ -1,88 +1,0 @@
-# Figma 轉測試案例SPEC
-
-## 核心目的
-此工具的核心目標是利用 AI 自動化分析 Figma 設計稿，並產出結構化的測試案例。主要包含四個關鍵環節：
-1. 輸入 Figma 網址：使用者提供一個 Figma 設計稿的公開網址作為輸入。
-2. AI 分析頁面內容：AI 會解析 Figma 檔案中每個畫面的具體內容，例如有哪些按鈕、表單、文字等元素。
-3. AI 分析頁面關聯：AI 會進一步分析不同頁面之間的流程關係與互動連結，例如點擊某個按鈕會跳轉到哪個頁面。
-4. 產出整合性測試案例：基於對頁面內容和流程的理解，AI 會自動生成一份完整的整合測試案例。
-
-## 操作方式
-定義了使用者與此工具的互動模式：
-1. CLI 介面：這是一個命令列工具 (Command Line Interface)，使用者將透過終端機下指令來操作。
-2. 輸入 Figma URL：主要的操作就是輸入一個 Figma 的網址。
-3. 輸出 Excel 檔案：執行成功後，工具會產出一個 Excel 格式的檔案，其中包含了所有生成出的測試案例。
-
-## Excel 測試案例格式
-這部分詳細定義了輸出 Excel 檔案中應包含的欄位，確保測試案例的標準化與完整性：
-1. 測案編號 (Test Case ID)：每個測試案例的唯一識別碼。
-2. 測試大項 (Test Suite)：測試案例所屬的主要功能模組或類別。
-3. 測試小項 (Test Section)：在主要功能模組下的次要分類。
-4. 優先等級 (Priority)：此測試案例的重要性，例如：高、中、低。
-5. 測試類型 (Test Category)：測試的種類，例如：功能測試、UI 測試、可用性測試等。
-6. 前置條件 (Precondition)：執行此測試前必須滿足的條件。
-7. 測試項目/步驟 (Test Step)：具體描述執行測試時需要操作的每一個步驟。
-8. 預期結果 (Expected Result)：執行完測試步驟後，預期應該看到的正確結果。
-
-# 專案的預計工作項目
-
-## 階段一：專案初始化與 CLI 框架建立
-
-1.  **環境設定**:
-    *   開發語言使用 Python
-    *   幫我建立專案資料夾結構
-    *   設定虛擬環境並安裝基礎套件。
-
-2.  **建立 CLI (命令列介面) 框架**:
-    *   使用 `argparse` 或 `click` 等套件建立一個基本的 CLI 應用程式。
-    *   使其能夠接收一個必要參數：`figma_url`。
-    *   加入 `--help` 或 `-h` 等輔助說明指令。
-
-## 階段二：Figma API 串接與資料獲取
-
-3.  **研究 Figma API**:
-    *   閱讀 Figma API 官方文件，了解如何進行身份驗證 (Authentication)。
-    *   找出獲取檔案 (File) 內容的相關 API 端點 (Endpoint)。
-    *   了解回傳的 JSON 資料結構，特別是關於頁面 (Pages)、畫框 (Frames)、節點 (Nodes) 和元件 (Components) 的部分。
-
-4.  **實作 Figma API 串接**:
-    *   撰寫程式碼，讓應用程式能使用使用者提供的 Figma Personal Access Token 進行驗證。
-    *   實作 API 呼叫，根據使用者輸入的 `figma_url` 獲取對應的 Figma 檔案 JSON 內容。
-    *   加入錯誤處理機制 (例如：無效的 URL、API Token 錯誤、網路問題)。
-
-## 階段三：AI 核心分析與測試案例生成
-
-5.  **資料前處理**:
-    *   解析從 Figma API 獲取的複雜 JSON 資料。
-    *   提取出對生成測試案例有用的關鍵資訊，例如：頁面名稱、可互動元件 (按鈕、輸入框)、元件之間的連結關係等。
-    *   將這些資訊整理成一個簡潔、結構化的格式，方便傳遞給 AI 模型。
-
-6.  **AI Prompt 設計與串接**:
-    *   設計一個強大的 Prompt (提示)，指示 AI 模型根據收到的結構化 Figma 資訊，執行以下任務：
-        *   分析每個頁面的功能。
-        *   推斷頁面之間的流程與使用者路徑。
-        *   根據分析結果，生成符合定義的 Excel 格式的測試案例。
-    *   串接一個大型語言模型 (LLM) 的 API (例如 Google AI API)。
-    *   將整理好的 Figma 資訊與設計好的 Prompt 發送給 AI 模型。
-
-## 階段四：產出 Excel 檔案
-
-7.  **處理 AI 回傳結果**:
-    *   接收 AI 模型回傳的測試案例內容 (可能是 JSON 或結構化文字)。
-    *   解析這些內容，並將其轉換為程式可處理的資料結構。
-
-8.  **生成 Excel 檔案**:
-    *   使用 `pandas` 或 `openpyxl` 等函式庫。
-    *   建立一個新的 Excel 工作簿 (Workbook)。
-    *   依照定義的欄位順序，寫入表頭 (Header)。
-    *   將從 AI 獲取的測試案例逐行寫入 Excel 表格中。
-    *   將檔案儲存為 `test_cases.xlsx` (或由使用者自訂檔名)。
-
-## 階段五：整合與優化
-
-9.  **整合所有模組**:
-    *   將 CLI 介面、Figma API 串接、AI 分析、Excel 生成等所有功能模組串連起來，確保端到端的流程順暢。
-
-10. **加入使用者說明與文件**:
-    *   撰寫一份 `README.md`，說明如何安裝、設定 (如何取得 Figma Token) 以及如何使用此工具。
-    *   完善 CLI 的輔助說明文字，引導使用者正確操作。
