+++ src/configs/global_adapter.py
@@ -1,0 +1,2 @@
+# ==== 通用設定 ====
+URL = ""
